Separate classes into correct files
