+++ headers/department.h
@@ -1,0 +1,11 @@
+#pragma once
+#include "course.h"
+#include "instructor.h"
+#include "room.h"
+#include <string>
+#include <vector>
+using namespace std;
+
+vector<Course> courseList;
+vector<Instructor> instructiorList;
+vector<RoomInfo> roomList;
