+++ x64/Debug/Course_Scheduling.log
@@ -1,0 +1,2 @@
+﻿  dayTime.cpp
+  Course_Scheduling.vcxproj -> C:\Users\henry\Documents\University of Alabama in Huntsville\CS 499-01\Course_Scheduling\x64\Debug\Course_Scheduling.exe
